refactor(maimai): plate handler 内联到 MessageHandler.summary section
按 review 要求调整结构 (#37 review 1/2/3/5)：
- 删 MaiMaiDx.PlateProgress.cs 独立文件，把 trigger + handler 塞进
  MessageHandler.cs 的 #region 汇总 / summary
- handler PlateProgress → Plate
- MarisaPluginDoc 缩成一句话；解析失败时 reply 加完整 PlateUsage 帮助
- SelectChartsForQuery / FormatError 改成 Plate handler 内 local function

渲染层暂未动 (DrawPlateProgress 还是 SkiaSharp)，下一个 commit 跟旧 sum
一起迁前端 (review 4)。
